Added oik-admin, bobbforms and bobbfunc for admin page
Decided to add the library files that would allow oik-lib to
display its simple admin page, rather than having to depend upon oik-bwtrace.

This allows oik-fum to be dependent upon oik-lib or the oik base plugin.
Not that there's any need for oik-fum when you have the oik base plugin.