Add require_file() method
To support Composer libraries with multiple files the require_file() method
will use $lib->path to locate the directory for the given file.
We assume that the file will be in the library that's loaded.
If not then a WP_Error will be returned.

Added some basic version checking.

This also caters for the required version being any ( * ).
When this is the case it doesn't matter what the current version is.
Semantic version checking is for the future.